Making all models serializable, added trail journals to be saved with trail, added text field hints for sleds and trails

+++ app/src/main/java/com/jcoapps/snowmobile_trail_maps/dao/TrailJournalsDao.java
@@ -40,9 +40,9 @@ public class TrailJournalsDao {
                 journal.setUpdatedAt(new Timestamp(Long.parseLong(journalsCursor.getString(2))));
                 journal.setEntryName(journalsCursor.getString(3));
                 journal.setMiles(journalsCursor.getDouble(4));
-                journal.setMaxSpeed(journalsCursor.getDouble(5));
-                journal.setMinSpeed(journalsCursor.getDouble(6));
-                journal.setAvgSpeed(journalsCursor.getDouble(7));
+                journal.setMaxSpeed(journalsCursor.getInt(5));
+                journal.setMinSpeed(journalsCursor.getInt(6));
+                journal.setAvgSpeed(journalsCursor.getInt(7));
 
                 /*
                 // TODO call TrailsDao
